feat(irs-api):[#470] Split SemanticsHubClient interface, implementation and stub in to separate files
